+++ tests/tasks/test_autotasks.py
@@ -8,8 +8,8 @@ from pathlib import Path
 import ray
 from flash.image import ImageClassificationData
 
-from gradsflow.core.model import Model
 from gradsflow.data.image import image_dataset_from_directory
+from gradsflow.model.model import Model
 from gradsflow.tasks import autotask
 
 warnings.filterwarnings("ignore")
